# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateListingTemplateRequest": {
        "description": "Save as listing template (``from_listing``) or Clone\n(``from_template``) -- exactly one. There is no blank creation (spec).",
        "properties": {
          "document": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          },
          "from_listing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Listing"
          },
          "from_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Template"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateListingTemplateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListingTemplateSummary": {
        "description": "One card on the Listing templates page (UI doc \u00a72).",
        "properties": {
          "batch_count": {
            "default": 0,
            "title": "Batch Count",
            "type": "integer"
          },
          "colour_count": {
            "title": "Colour Count",
            "type": "integer"
          },
          "design_minimum": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PixelSize"
              },
              {
                "type": "null"
              }
            ]
          },
          "garment": {
            "title": "Garment",
            "type": "string"
          },
          "media": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TemplateMediaEntry"
                },
                {
                  "type": "string"
                }
              ]
            },
            "title": "Media",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pricing_plan_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricing Plan Name"
          }
        },
        "required": [
          "name",
          "garment",
          "colour_count",
          "pricing_plan_name",
          "media"
        ],
        "title": "ListingTemplateSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "etsy-listings",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/listing-templates": {
      "get": {
        "description": "Every listing template, as its card on the Listing templates page shows\nit (UI doc \u00a72). A template whose file will not load is left out rather\nthan failing the page; it can only get that way by a hand edit.",
        "operationId": "list_listing_templates_api_listing_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ListingTemplateSummary"
                  },
                  "title": "Response List Listing Templates Api Listing Templates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Listing Templates",
        "tags": [
          "listing-templates"
        ]
      },
      "post": {
        "description": "Naming the draft is what writes it (UI doc \u00a71). The name is checked\nfirst, so a bad one is a 400 before any source is read.",
        "operationId": "create_listing_template_api_listing_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingTemplateSaveResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Listing Template",
        "tags": [
          "listing-templates"
        ]
      }
    },
    "/api/listing-templates/draft": {
      "get": {
        "description": "The template Save as listing template or Clone would write, written\nnowhere: what the *name it* page shows before the seller commits a name\n(UI doc \u00a71).",
        "operationId": "listing_template_draft_api_listing_templates_draft_get",
        "parameters": [
          {
            "in": "query",
            "name": "from_listing",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Listing"
            }
          },
          {
            "in": "query",
            "name": "from_template",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Template"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingTemplateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing Template Draft",
        "tags": [
          "listing-templates"
        ]
      }
    },
    "/api/listing-templates/{name}": {
      "delete": {
        "description": "Allowed whatever was made from it: batches keep their own frozen copy,\nand a listing never had a link to it (spec, *Completeness and editing*).",
        "operationId": "delete_listing_template_api_listing_templates__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Listing Template",
        "tags": [
          "listing-templates"
        ]
      },
      "get": {
        "operationId": "get_listing_template_api_listing_templates__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingTemplateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Listing Template",
        "tags": [
          "listing-templates"
        ]
      },
      "put": {
        "description": "template completeness: write the whole document only if it is complete. A malformed or\nincomplete one leaves ``template.yaml`` byte-for-byte as it was, so the\nserver always holds the last complete version; the listing-template editor\n(PR 6) keeps the unsaved values on its side.",
        "operationId": "put_listing_template_api_listing_templates__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingTemplateSaveResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Listing Template",
        "tags": [
          "listing-templates"
        ]
      }
    },
    "/api/listing-templates/{name}/rename": {
      "post": {
        "description": "Move a listing template, whole, to a new name -- double-click the\nname, exactly as a listing (UI doc \u00a73). Its name is its directory,\nso the rename is a directory move and its ``./`` refs, which name that\ndirectory, need no rewrite. A taken name is a 409 and never suffixed\n(spec, *Storage and identity*).\n\nStaging sessions and batch records made from it follow it by name --\nthe name, not the frozen copy they each keep, which is untouched.\nThat name is the card's *Used by N batches* and the staging page's\n*Using X*, the seller's link between a template and its batches. It\nis rewritten after the move, under each record's lock; a crash between\nthe two leaves records naming a template that is gone, which is what a\ndelete leaves too, and harms nothing a batch needs.",
        "operationId": "rename_listing_template_api_listing_templates__name__rename_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingTemplateDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Listing Template",
        "tags": [
          "listing-templates"
        ]
      }
    },
...
    "/api/listings": {
      "get": {
        "description": "Every listing, with its status resolved in **one** Etsy round trip for\nthe whole table rather than one per row -- which is why the live fact is\ngathered here and handed down rather than looked up per listing.",
        "operationId": "list_listings_api_listings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ListingSummary"
                  },
                  "title": "Response List Listings Api Listings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Listings",
        "tags": [
          "listings"
        ]
      },
      "post": {
        "description": "Naming a listing is what creates it.\n\nMirrors PATCH exactly, one level up: a document that fails\n``Listing.model_validate`` is a **200** carrying ``field_errors`` with\nnothing written, so the editor never has to branch on a status code to show\ninline validation. The two things a *name* can be wrong about keep their\nstatus codes instead -- not a single path segment is the 400 `_segment`\nraises through `listing_file`, and already taken is a 409.\n\nThat 409 tests the **directory**, not ``listing.yaml``: a `listings/{name}/`\nleft behind with a `state.lock.json` and no document would otherwise be\nwritten into, and the new listing would inherit another one's\n``etsy_listing_id``.",
        "operationId": "create_listing_api_listings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Listing",
        "tags": [
          "listings"
        ]
      }
    },
...
    "/api/listings/{name}/rename": {
      "post": {
        "description": "Move a listing, whole, to a new name.\n\nA listing's identity is its directory name, so the rename is a\ndirectory move: ``listing.yaml``, ``state.lock.json`` and Phase 4's\ngenerated copy travel together, and `.cache/renders/{name}/` moves with them\nbecause the render cache is keyed by listing name too -- left behind it\nwould orphan a tree nothing deletes and cost a full re-render. The market\nsnapshot and the cached AI proposal move for the same reason, and\nevery batch row naming the listing follows it, so the batch\nsummary opens the new name.\n\nThe lockfile's ``outputs`` keys still spell the old path afterwards, and are\nleft that way deliberately: nothing reads them, they become true again at\nthe next apply, and rewriting them here would breach \"only the lockfile\nmerges a lockfile\".\n\nPOST rather than PUT: the body is neither the listing nor its new\nrepresentation, and a second call 404s. `POST /api/templates/{name}/kind` is\nthe same shape.",
        "operationId": "rename_listing_api_listings__name__rename_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Listing",
        "tags": [
          "listings"
        ]
      }
    },
...
  }
}